refactor: merge NotchCompanionView extensions from 4 to 3 files
- Merge +Controls extension into main NotchCompanionView.swift

- Keep +StandardViews and +InsideNotch as separate extension files

- All files stay under swiftlint 500-line limit

- Reduces file fragmentation by consolidating control code
